+++ Lib/fontbakery/profiles/fontbureau.py
@@ -3,6 +3,7 @@ PROFILE = {
     "pending_review": [
         "opentype:weight_class_fvar",
         "opentype:slant_direction",
+        "family/control_chars",
         "fvar_name_entries",
         "fontdata_namecheck",
         "glyf_nested_components",