+++ Lib/fontbakery/profiles/microsoft.py
@@ -14,6 +14,7 @@ PROFILE = {
         "designspace_has_consistent_groups",
         "designspace_has_default_master",
         "designspace_has_sources",
+        "family/control_chars",
         "fvar_name_entries",
         "fontdata_namecheck",
         "glyf_nested_components",