+++ Lib/fontbakery/profiles/fontbureau.py
@@ -17,6 +17,7 @@ PROFILE = {
         "no_mac_entries",
         "render_own_name",
         "stylisticset_description",
+        "typographic_family_name",
         "vtt_volt_data",  # very similar to vttclean, may be a good idea to merge them.
         "vttclean",
     ],