+++ Lib/fontbakery/profiles/adobefonts.py
@@ -73,6 +73,7 @@ PROFILE = {
         "typoascender_exceeds_Agrave",
         "typographic_family_name",
         "unwanted_tables",
+        "varfont/consistent_axes",
         "vtt_volt_data",  # very similar to vttclean, may be a good idea to merge them.
         "vttclean",
     ],