+++ Lib/fontbakery/profiles/universal.py
@@ -75,6 +75,7 @@ PROFILE = {
             "unwanted_aat_tables",
             "unwanted_tables",
             "valid_glyphnames",
+            "varfont/consistent_axes",
             "vtt_volt_data",  # very similar to vttclean, may be a good idea to merge them.
             "vttclean",
             "whitespace_glyphnames",