+++ Lib/fontbakery/profiles/adobefonts.py
@@ -52,6 +52,7 @@ PROFILE = {
         "color_cpal_brightness",
         "colorfont_tables",
         "contour_count",  # was temporarily removed
+        "empty_glyph_on_gid1_for_colrv0",
         "family/control_chars",
         "file_size",
         "fontdata_namecheck",