+++ Lib/fontbakery/profiles/fontbureau.py
@@ -30,6 +30,7 @@ PROFILE = {
         "typographic_family_name",
         "varfont/consistent_axes",
         "varfont/duplexed_axis_reflow",
+        "varfont/instances_in_order",
         "vtt_volt_data",  # very similar to vttclean, may be a good idea to merge them.
         "vttclean",
     ],