+++ Lib/fontbakery/profiles/microsoft.py
@@ -49,6 +49,7 @@ PROFILE = {
         "unwanted_aat_tables",
         "varfont/consistent_axes",
         "varfont/duplexed_axis_reflow",
+        "varfont/instances_in_order",
         "vttclean",  # very similar to vtt_volt_data, may be a good idea to merge them.
     ],
     "sections": {