+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Below are the noteworthy changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 ##  Upcoming release: 0.12.11 (2024-Aug-??)
-  - ...
+### Migration of checks
+#### Moved to the Universal profile
+  - **[com.google.fonts/check/aat]:** from the Google Fonts profile. (PR #4820)
 
 
 ##  0.12.10 (2024-Aug-14)